Extract duplicate decodeBasic logic into shared BasicAuthUtils
Extracted duplicate Base64 Basic Auth decoding logic from
LegacyAuthentication.kt and PostService.kt into a new shared utility:
org.bibsonomy.api.security.BasicAuthUtils

Changes:
- Created BasicAuthUtils.decode(header) for consistent auth parsing
- Moved BASIC_PREFIX constant to BasicAuthUtils
- Replaced both decodeBasic implementations with BasicAuthUtils.decode
- Unified error handling: both now throw BadCredentialsException
- Removed unused imports (Base64, StandardCharsets) from PostService

Benefits:
- Single source of truth for Basic Auth decoding logic
- Consistent error messages and exception types
- Easier to maintain and test
- DRY (Don't Repeat Yourself) principle
